test(control-flow): isolate localized error fixtures
