+++ tests/CMakeLists.txt
@@ -5531,6 +5531,7 @@ copperfin_add_test(test_prg_engine_control_flow cf_xbase_runtime
     test_prg_engine_control_flow_error_handler_commands.cpp
     test_prg_engine_control_flow_bare_null.cpp
     test_prg_engine_control_flow_error_handling_and_faults.cpp
+    test_prg_engine_control_flow_localized_errors.cpp
     test_prg_engine_control_flow_fault_recovery.cpp
     test_prg_engine_control_flow_runtime_guardrails_and_lifecycle.cpp
     test_prg_engine_control_flow_routine_calls_parameters.cpp
